reorganized file structure for components, refactored member page to be modularized and contain helpers
