+++ client/src/components/shared/footer.js
@@ -1,0 +1,16 @@
+import React from 'react';
+
+const FooterComponent = () => {
+    return(
+        <div className="footer">
+            <p>Made with <span><i className="fa fa-heart"></i>
+            <a target="_blank"
+               href="https://github.com/kapitol-app/kapitol-platform"
+               rel="noopener noreferrer">by the people
+            </a>,
+            for the people</span></p>
+        </div>
+    )
+}
+
+export default FooterComponent;
